+++ src/ContaoManager/Plugin.php
@@ -12,9 +12,12 @@ use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
+use Contao\ManagerPlugin\Config\ContainerBuilder;
+use Contao\ManagerPlugin\Config\ExtensionPluginInterface;
 use HeimrichHannot\CategoriesBundle\CategoriesBundle;
+use HeimrichHannot\UtilsBundle\Container\ContainerUtil;
 
-class Plugin implements BundlePluginInterface
+class Plugin implements BundlePluginInterface, ExtensionPluginInterface
 {
     /**
      * {@inheritdoc}
@@ -23,7 +26,29 @@ class Plugin implements BundlePluginInterface
     {
         return [
             BundleConfig::create(CategoriesBundle::class)
-                ->setLoadAfter([ContaoCoreBundle::class]),
+                ->setLoadAfter([
+                    ContaoCoreBundle::class,
+                    'HeimrichHannot\FilterBundle\HeimrichHannotContaoFilterBundle'
+                ]),
         ];
+    }
+
+    /**
+     * Allows a plugin to override extension configuration.
+     *
+     * @param string $extensionName
+     * @param array $extensionConfigs
+     * @param ContainerBuilder $container
+     *
+     * @return array
+     */
+    public function getExtensionConfig($extensionName, array $extensionConfigs, ContainerBuilder $container)
+    {
+        return ContainerUtil::mergeConfigFile(
+            'huh_filter',
+            $extensionName,
+            $extensionConfigs,
+            __DIR__ .'/../Resources/config/huh_filter.yml'
+        );
     }
 }