+++ src/Resources/contao/languages/de/tl_filter_config_element.php
@@ -1,0 +1,6 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_filter_config_element'];
+
+$lang['parentCategories'] = ["Elternkategorien", "Wählen Sie hier die Elternkategorien der Kategorien aus, welche im Filter gewählt werden können."];
+$lang['reference']['type'][\HeimrichHannot\CategoriesBundle\Filter\Type\CategoryChoiceType::TYPE] = "Kategorien";