+++ src/Resources/contao/languages/en/tl_filter_config_element.php
@@ -1,0 +1,6 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_filter_config_element'];
+
+$lang['parentCategories'] = ["Parent categories", "Choose parent categories of the categories, which should be available in the filter."];
+$lang['reference']['type'][\HeimrichHannot\CategoriesBundle\Filter\Type\CategoryChoiceType::TYPE] = "Categories";